refactor: useProductSummary props 이름 변경
