+++ src/page/Order/hooks/useOrderForm.ts
@@ -1,26 +1,26 @@
-import { useParams } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import type { OrderInfoValues } from '@/types';
-import useProductSummary from './useProductSummary';
+import useProductSummary from '../../../hooks/useProductSummary';
 import useSubmitOrder from './useSubmitOrder';
 import showOrderSuccessAlert from '../utils/showOrderSuccessAlert';
 import useInitOrderForm from './useInitOrderForm';
+import { useParamsIndex } from '@/hooks/useParamsIndex';
 
 const useOrderForm = () => {
-  const { id } = useParams<{ id: string }>();
-  const productSummaryData = useProductSummary(id as string);
+  const index = useParamsIndex();
+  const productSummaryData = useProductSummary(index);
   const orderForm = useInitOrderForm();
   const orderData = orderForm.getValues();
   const price = orderData.receiverInfos.length * (productSummaryData?.price || 0);
   const { mutate: order } = useSubmitOrder();
 
   const onSubmit = async (formData: OrderInfoValues) => {
-    if (!id) return;
+    if (!index) return;
     if (orderData.receiverInfos.length === 0) {
       toast('받는 사람이 없습니다');
       return;
     }
-    order({ orderData: formData, id });
+    order({ orderData: formData, index });
     showOrderSuccessAlert({ productSummaryData, formData });
   };
